refactor: imageSetup API 레이어 구조 개선
- apis/와 hooks/queries 폴더를 apis/로 통합
- API 함수와 Tanstack Query hooks를 단일 파일로 결합
