--- src/pages/imageSetup/apis/FloorPlan.ts
@@ -1,11 +1,0 @@
-import { type FloorPlanResponse } from '../types/apis/floorPlan';
-import { HTTPMethod, type RequestConfig, request } from '@/shared/apis/request';
-
-export const getFloorPlan = async (): Promise<FloorPlanResponse['data']> => {
-  const config: RequestConfig = {
-    method: HTTPMethod.GET,
-    url: '/api/v1/house-templates',
-  };
-
-  return await request<FloorPlanResponse['data']>(config);
-};
