+++ src/pages/imageSetup/components/steps/interiorStyle/MoodBoard.tsx
@@ -16,7 +16,7 @@ import {
   MOOD_BOARD_CONSTANTS,
   type MoodBoardImageItem,
 } from '@/pages/imageSetup/types/apis/interiorStyle';
-import { useMoodBoardQuery } from '@/pages/imageSetup/hooks/queries/useInteriorStyleApi';
+import { useMoodBoardQuery } from '@/pages/imageSetup/api/interiorStyleApi';
 import CardImage from '@/shared/components/card/cardImage/CardImage';
 import SkeletonCardImage from '@/shared/components/card/cardImage/SkeletonCardImage';
 import { useErrorHandler } from '@/shared/hooks/useErrorHandler';
