+++ src/pages/imageSetup/hooks/useActivityInfo.ts
@@ -4,8 +4,8 @@ import {
   MAIN_ACTIVITY_OPTIONS,
   MAIN_ACTIVITY_VALIDATION,
   type ImageSetupSteps,
-} from '../../types/funnel';
-import { useFunnelStore } from '../../stores/useFunnelStore';
+} from '../types/funnel';
+import { useFunnelStore } from '../stores/useFunnelStore';
 import type { GenerateImageRequest } from '@/pages/generate/types/GenerateType';
 import { ROUTES } from '@/routes/paths';
 import { useCreditGuard } from '@/shared/hooks/useCreditGuard';
