+++ src/pages/imageSetup/hooks/useInteriorStyle.ts
@@ -1,10 +1,7 @@
 import { useCallback, useEffect, useState } from 'react';
-import { MAX_MOOD_BOARD_SELECTION } from '../../constants/interiorStyle';
-import { useFunnelStore } from '../../stores/useFunnelStore';
-import type {
-  CompletedInteriorStyle,
-  ImageSetupSteps,
-} from '../../types/funnel';
+import { MAX_MOOD_BOARD_SELECTION } from '../constants/interiorStyle';
+import { useFunnelStore } from '../stores/useFunnelStore';
+import type { CompletedInteriorStyle, ImageSetupSteps } from '../types/funnel';
 
 export const useInteriorStyle = (
   context: ImageSetupSteps['InteriorStyle'],
